attempt to reduce code duplication in code quizes
