+++ app/src/main/java/org/stepik/android/view/injection/step/StepComponent.kt
@@ -7,10 +7,8 @@ import org.stepik.android.view.injection.step_quiz.StepQuizModule
 import org.stepik.android.view.injection.step_quiz.StepQuizPresentationModule
 import org.stepik.android.view.injection.submission.SubmissionDataModule
 import org.stepik.android.view.step.ui.fragment.StepFragment
+import org.stepik.android.view.step_quiz.ui.fragment.DefaultStepQuizFragment
 import org.stepik.android.view.step_quiz_code.ui.fragment.CodeStepQuizFragment
-import org.stepik.android.view.step_quiz_choice.ui.fragment.ChoiceStepQuizFragment
-import org.stepik.android.view.step_quiz_sorting.ui.fragment.SortingStepQuizFragment
-import org.stepik.android.view.step_quiz_text.ui.fragment.TextStepQuizFragment
 
 @Subcomponent(modules = [
     StepModule::class,
@@ -29,8 +27,6 @@ interface StepComponent {
 
     fun inject(stepFragment: StepFragment)
 
-    fun inject(choiceStepQuizFragment: ChoiceStepQuizFragment)
     fun inject(codeStepQuizFragment: CodeStepQuizFragment)
-    fun inject(textStepQuizFragment: TextStepQuizFragment)
-    fun inject(sortingStepQuizFragment: SortingStepQuizFragment)
+    fun inject(defaultStepQuizFragment: DefaultStepQuizFragment)
 }
