+++ app/src/main/java/org/stepik/android/view/step_quiz_choice/ui/fragment/ChoiceStepQuizFragment.kt
@@ -1,36 +1,17 @@
 package org.stepik.android.view.step_quiz_choice.ui.fragment
 
-import android.arch.lifecycle.ViewModelProvider
-import android.arch.lifecycle.ViewModelProviders
-import android.os.Bundle
-import android.support.design.widget.Snackbar
 import android.support.v4.app.Fragment
-import android.support.v4.content.ContextCompat
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import kotlinx.android.synthetic.main.error_no_connection_with_button_small.view.*
-import kotlinx.android.synthetic.main.fragment_step_quiz.*
 import kotlinx.android.synthetic.main.layout_step_quiz_choice.*
-import kotlinx.android.synthetic.main.view_step_quiz_submit_button.*
 import org.stepic.droid.R
-import org.stepic.droid.base.App
-import org.stepic.droid.core.ScreenManager
-import org.stepic.droid.fonts.FontsProvider
 import org.stepic.droid.persistence.model.StepPersistentWrapper
-import org.stepic.droid.ui.listeners.NextMoveable
-import org.stepic.droid.util.argument
-import org.stepic.droid.util.setTextColor
 import org.stepik.android.domain.lesson.model.LessonData
-import org.stepik.android.presentation.step_quiz.StepQuizPresenter
 import org.stepik.android.presentation.step_quiz.StepQuizView
-import org.stepik.android.view.step_quiz.ui.delegate.StepQuizDelegate
-import org.stepik.android.view.step_quiz.ui.delegate.StepQuizFeedbackBlocksDelegate
+import org.stepik.android.view.step_quiz.ui.delegate.StepQuizFormDelegate
+import org.stepik.android.view.step_quiz.ui.fragment.DefaultStepQuizFragment
 import org.stepik.android.view.step_quiz_choice.ui.delegate.ChoiceStepQuizFormDelegate
-import org.stepik.android.view.ui.delegate.ViewStateDelegate
-import javax.inject.Inject
 
-class ChoiceStepQuizFragment : Fragment(), StepQuizView {
+class ChoiceStepQuizFragment : DefaultStepQuizFragment(), StepQuizView {
     companion object {
         fun newInstance(stepPersistentWrapper: StepPersistentWrapper, lessonData: LessonData): Fragment =
             ChoiceStepQuizFragment()
@@ -40,99 +21,12 @@ class ChoiceStepQuizFragment : Fragment(), StepQuizView {
                 }
     }
 
-    @Inject
-    internal lateinit var viewModelFactory: ViewModelProvider.Factory
+    override val quizLayoutRes: Int =
+        R.layout.layout_step_quiz_choice
 
-    @Inject
-    internal lateinit var fontsProvider: FontsProvider
+    override val quizViews: Array<View>
+        get() = arrayOf(choicesRecycler)
 
-    @Inject
-    internal lateinit var screenManager: ScreenManager
-
-    private lateinit var presenter: StepQuizPresenter
-
-    private var lessonData: LessonData by argument()
-    private var stepWrapper: StepPersistentWrapper by argument()
-
-    private lateinit var viewStateDelegate: ViewStateDelegate<StepQuizView.State>
-    private lateinit var stepQuizDelegate: StepQuizDelegate
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        injectComponent()
-
-        presenter = ViewModelProviders.of(this, viewModelFactory).get(StepQuizPresenter::class.java)
-        presenter.onStepData(stepWrapper, lessonData)
-    }
-
-    private fun injectComponent() {
-        App.component()
-            .stepComponentBuilder()
-            .build()
-            .inject(this)
-    }
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
-        (inflater.inflate(R.layout.fragment_step_quiz, container, false) as ViewGroup)
-            .apply {
-                addView(inflater.inflate(R.layout.layout_step_quiz_choice, this, false))
-            }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        viewStateDelegate = ViewStateDelegate()
-        viewStateDelegate.addState<StepQuizView.State.Idle>()
-        viewStateDelegate.addState<StepQuizView.State.Loading>(stepQuizProgress)
-        viewStateDelegate.addState<StepQuizView.State.AttemptLoaded>(stepQuizDiscountingPolicy, stepQuizFeedbackBlocks, choicesRecycler, stepQuizDescription, stepQuizActionContainer)
-        viewStateDelegate.addState<StepQuizView.State.NetworkError>(stepQuizNetworkError)
-
-        stepQuizNetworkError.tryAgain.setOnClickListener { presenter.onStepData(stepWrapper, lessonData, forceUpdate = true) }
-
-        stepQuizDelegate =
-            StepQuizDelegate(
-                step = stepWrapper.step,
-                lessonData = lessonData,
-                stepQuizFormDelegate = ChoiceStepQuizFormDelegate(view, fontsProvider),
-                stepQuizFeedbackBlocksDelegate = StepQuizFeedbackBlocksDelegate(
-                    stepQuizFeedbackBlocks,
-                    fontsProvider,
-                    stepWrapper.step.actions?.doReview != null,
-                    { screenManager.openStepInWeb(requireContext(), stepWrapper.step) }
-                ),
-                stepQuizActionButton = stepQuizAction,
-                stepRetryButton = stepQuizRetry,
-                stepQuizDiscountingPolicy = stepQuizDiscountingPolicy,
-                stepQuizPresenter = presenter
-            ) {
-                (parentFragment as? NextMoveable)?.moveNext()
-            }
-    }
-
-    override fun onStart() {
-        super.onStart()
-        presenter.attachView(this)
-    }
-
-    override fun onStop() {
-        presenter.detachView(this)
-        stepQuizDelegate.syncReplyState()
-        super.onStop()
-    }
-
-    override fun setState(state: StepQuizView.State) {
-        viewStateDelegate.switchState(state)
-        if (state is StepQuizView.State.AttemptLoaded) {
-            stepQuizDelegate.setState(state)
-        }
-    }
-
-    override fun showNetworkError() {
-        val view = view ?: return
-
-        Snackbar
-            .make(view, R.string.no_connection, Snackbar.LENGTH_SHORT)
-            .setTextColor(ContextCompat.getColor(requireContext(), R.color.white))
-            .show()
-    }
+    override fun createStepQuizFormDelegate(view: View): StepQuizFormDelegate =
+        ChoiceStepQuizFormDelegate(view, fontsProvider)
 }
